segregated ProjectTest and CreateProjectTest
